# services/Diten.PlanningService/contracts/drafts/demand-v2.openapi.yaml
openapi: 3.1.0
info:
  title: MOD-0188 Demand v2 producer contract (DRAFT)
  version: 2.0.0-draft.1
  description: >-
    MOD-0188-owned review draft for immutable gross-demand revisions. It is not
    frozen, published, routed, or evidence of a working producer. Demand v1
    remains unchanged. MRP period choice, order consumption, MRP run stops and
    reconciliation belong to MOD-0189, not this producer API.
  x-status: DRAFT
  x-owner: MOD-0188
  x-service: Diten.PlanningService
  x-central-freeze-required: true
  x-event-schema: ./demand-v2.events.schema.json
servers:
  - url: /api/v2/demand
    description: Candidate service path; Gateway publication requires central approval.
security:
  - bearerAuth: []
...
paths:
...
  /revisions/{revisionId}/rollback-drafts:
    post:
      operationId: createRollbackDemandDraft
      summary: Copy a Published or Superseded manual snapshot into a new Draft
      description: >-
        Creates a distinct revision in the same locked planning cycle. Only
        selected snapshot weeks and recorded exclusions are copied; original
        manual-week and exclusion actor, reason and time remain lineage. The
        requesting actor becomes the new preparer. Invalidated or unverifiable
        source and method-origin weeks without sufficient provenance are not
        copied. The source revision, checksum, status and audit are unchanged.
        This does not approve or publish the new Draft. Live scope authority
        remains unconfigured and fails closed outside explicit test fixtures.
      x-required-permission: demand.drafts.create
      parameters:
        - { $ref: '#/components/parameters/RevisionId' }
        - { $ref: '#/components/parameters/SelectedLegalEntityHint' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RollbackDraftCommand' } } }
      responses:
        '201': { description: New Draft in Response envelope, content: { application/json: { schema: { $ref: '#/components/schemas/ManualDraftEnvelope' } } } }
        '200': { description: Idempotent replay of the original Draft, without a second effect, content: { application/json: { schema: { $ref: '#/components/schemas/ManualDraftEnvelope' } } } }
        '400': { $ref: '#/components/responses/DraftFailure' }
        '401': { description: Unauthenticated; JWT middleware may return no body. }
        '403': { $ref: '#/components/responses/DraftFailure' }
        '404': { $ref: '#/components/responses/DraftFailure' }
        '409': { $ref: '#/components/responses/DraftFailure' }
        '422': { $ref: '#/components/responses/DraftFailure' }
        '503': { $ref: '#/components/responses/DraftFailure' }
  /revisions/{revisionId}/publish:
    post:
      operationId: publishDemandRevision
      summary: Publish a valid Approved revision
      description: >-
        Separation of duties, all-series and part/row/checksum validation,
        and one current Published baseline per Tenant x LegalEntity x
        planningPeriodKey are mandatory. A successful new publication
        atomically supersedes the previous current baseline and emits one
        outbox publication event. Whether the prior revision also emits a
        Superseded event or exposes an equivalent verifiable status change is
        reserved for central review; its authoritative state and stateVersion
        must change either way. Failure creates no Published state or event.
      x-required-permission: demand.plans.publish
      parameters:
        - { $ref: '#/components/parameters/RevisionId' }
        - { $ref: '#/components/parameters/SelectedLegalEntityHint' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublishCommand' }
      responses:
        '200': { description: Publication receipt, not an authoritative current-status read; replay has no second effect., content: { application/json: { schema: { $ref: '#/components/schemas/RevisionCommandEnvelope' } } } }
        '400': { $ref: '#/components/responses/RevisionCommandFailure' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/RevisionCommandFailure' }
        '404': { $ref: '#/components/responses/RevisionCommandFailure' }
        '409': { $ref: '#/components/responses/RevisionCommandFailure' }
        '422': { $ref: '#/components/responses/RevisionCommandFailure' }
        '503': { $ref: '#/components/responses/RevisionCommandFailure' }
  /revisions/{revisionId}/invalidate:
    post:
      operationId: invalidateDemandRevision
      summary: Protectively invalidate a materially erroneous revision
      description: >-
        Only Published or Superseded can become Invalidated. A qualified
        planning/quality manager provides reason and evidence of material MRP
        effect. Natural forecast deviation alone is insufficient. This does
        not cancel formal downstream transactions or restore an older revision.
      x-required-permission: demand.plans.invalidate
      parameters:
        - { $ref: '#/components/parameters/RevisionId' }
        - { $ref: '#/components/parameters/SelectedLegalEntityHint' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InvalidateCommand' }
      responses:
        '200': { description: Invalidation receipt, not an authoritative current-status read; replay has no second effect., content: { application/json: { schema: { $ref: '#/components/schemas/RevisionCommandEnvelope' } } } }
        '400': { $ref: '#/components/responses/RevisionCommandFailure' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/RevisionCommandFailure' }
        '404': { $ref: '#/components/responses/RevisionCommandFailure' }
        '409': { $ref: '#/components/responses/RevisionCommandFailure' }
        '422': { $ref: '#/components/responses/RevisionCommandFailure' }
        '503': { $ref: '#/components/responses/RevisionCommandFailure' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    SelectedLegalEntityHint:
      name: X-Legal-Entity-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: Untrusted single-company selection hint; current server-side assignment and exclusions must still be verified.
    RevisionId: { name: revisionId, in: path, required: true, schema: { type: string, format: uuid } }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1 }
      description: Same key and content repeat one effect; same key with different content conflicts.
  responses:
    RevisionCommandFailure:
      description: No command effect. Policy denial may have no body; handler failures use Response envelope.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ReadEnvelope' }
...
    DraftFailure:
      description: >-
        Response envelope with no successful data. 403 means missing Demand
        permission; 404 hides an out-of-scope existing record; 409 means
        stale version, candidate slot or idempotency conflict; 422 is invalid
        manual content; 503 means scope, reference, storage or audit cannot
        be verified. The current nullable assignment seam cannot safely
        distinguish same-tenant denial from cross-tenant selection.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ManualDraftEnvelope' }
    Problem:
      description: >-
        Error, never partial success. 400 means malformed request; 401 means
        unauthenticated, 403 missing
        permission, 404 hides other-Tenant/out-of-scope records, 409 means stale
        state or invalidated consumption, 422 invalid business input, and 503
        unavailable authoritative scope/status/integrity or durable audit.
        Exact machine-readable error codes require central review.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    RevisionCommandEnvelope:
      allOf:
        - { $ref: '#/components/schemas/ReadEnvelope' }
        - type: object
          required: [data]
          properties: { data: { $ref: '#/components/schemas/RevisionCommandReceipt' } }
    RevisionCommandReceipt:
      type: object
      required: [revisionId, outcome]
      properties:
        revisionId: { type: string, format: uuid }
        outcome: { type: string, enum: [Published, Invalidated, Replayed] }
        stateVersion: { type: [integer, 'null'], description: Returned when available; replay never implies that this is still the current state. }
        checksum: { type: [string, 'null'], description: Returned for a new publication when available. }
...
    ReadEnvelope:
      type: object
      required: [statusCode, isSuccessful, errors]
      properties:
        statusCode: { type: integer }
        isSuccessful: { type: boolean }
        errors: { type: array, items: { type: string } }
        data: { type: [object, 'null'], description: Null on a failed read; success envelopes require a typed object. }
...
    ManualDraftEnvelope:
      type: object
      required: [statusCode, isSuccessful, errors]
      properties:
        statusCode: { type: integer }
        isSuccessful: { type: boolean }
        errors: { type: array, items: { type: string } }
        data: { $ref: '#/components/schemas/ManualDraftView' }
    ManualDraftView:
      type: object
      description: >-
        Draft-only mutable candidate, not a Published immutable manifest.
        Each selected series contains all 52 locked cycle weeks, source and
        per-week manual provenance; changes retain old/new values and actor/time.
        The current API sends enum values as numeric JSON pending central
        wire-format review. No method output is relabeled manual.
      required: [revisionId, planningCycleId, tenantId, legalEntityId, planningPeriodKey, asOfDate, calendarId, calendarVersion, timeZoneId, horizonStart, horizonEnd, state, contentVersion, stateVersion, createdBy, createdAt, creationReason, series, changes, exclusions, replayed]
      properties:
        revisionId: { type: string, format: uuid }
        planningCycleId: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        legalEntityId: { type: string, format: uuid }
        planningPeriodKey: { type: string, format: date }
        asOfDate: { type: string, format: date }
        calendarId: { type: string }
        calendarVersion: { type: string }
        timeZoneId: { type: string }
        horizonStart: { type: string, format: date }
        horizonEnd: { type: string, format: date }
        state: { type: integer, enum: [0, 1, 2], description: 'Current Draft/InReview/Approved candidate values; not the consumer RevisionState representation.' }
        contentVersion: { type: integer, minimum: 0 }
        stateVersion: { type: integer, minimum: 0 }
        createdBy: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        creationReason: { type: string }
        reviewedBy: { type: [string, 'null'], format: uuid }
        reviewedAt: { type: [string, 'null'], format: date-time }
        reviewReason: { type: [string, 'null'] }
        series: { type: array, minItems: 1, items: { $ref: '#/components/schemas/ManualDraftSeriesView' } }
        changes: { type: array, items: { $ref: '#/components/schemas/ManualDraftChange' } }
        exclusions: { type: array, items: { $ref: '#/components/schemas/ManualDraftSeriesExclusion' } }
        replayed: { type: boolean }
    ManualDraftSeriesView:
      type: object
      required: [skuId, warehouseId, baseUomId, weeks, exclusion]
      properties:
        skuId: { type: string, format: uuid }
        warehouseId: { type: string }
        baseUomId: { type: string, description: Server-verified SKU base unit; never taken from create payload. }
        exclusion: { oneOf: [{ type: 'null' }, { $ref: '#/components/schemas/ManualDraftSeriesExclusion' }] }
        weeks:
          type: array
          minItems: 52
          maxItems: 52
          items: { $ref: '#/components/schemas/ManualDraftWeekView' }
    ManualDraftWeekView:
      type: object
      required: [number, weekStart, weekEnd, valueKind, source, manualReason, manualActorId, manualAt]
      properties:
        number: { type: integer, minimum: 1, maximum: 52 }
        weekStart: { type: string, format: date }
        weekEnd: { type: string, format: date }
        valueKind: { $ref: '#/components/schemas/ManualWeekInput/properties/valueKind' }
        quantity: { type: [number, 'null'] }
        source: { type: integer, enum: [0, 1], description: '0 Manual, 1 Method; this manual Draft slice produces 0 only.' }
        manualReason: { type: string }
        manualActorId: { type: string, format: uuid }
        manualAt: { type: string, format: date-time }
    ManualDraftChange:
      type: object
      required: [skuId, warehouseId, weekNumber, oldKind, newKind, actorId, occurredAt, reason, requestKey]
      properties:
        skuId: { type: string, format: uuid }
        warehouseId: { type: string }
        weekNumber: { type: integer, minimum: 1, maximum: 52 }
        oldKind: { $ref: '#/components/schemas/ManualWeekInput/properties/valueKind' }
        oldQuantity: { type: [number, 'null'] }
        newKind: { $ref: '#/components/schemas/ManualWeekInput/properties/valueKind' }
        newQuantity: { type: [number, 'null'] }
        actorId: { type: string, format: uuid }
        occurredAt: { type: string, format: date-time }
        reason: { type: string }
        requestKey: { type: string }
    ManualDraftSeriesExclusion:
      type: object
      required: [skuId, warehouseId, actorId, occurredAt, reason, requestKey, previousState, newState]
      properties:
        skuId: { type: string, format: uuid }
        warehouseId: { type: string }
        actorId: { type: string, format: uuid }
        occurredAt: { type: string, format: date-time }
        reason: { type: string, minLength: 1 }
        requestKey: { type: string }
        previousState: { type: integer, enum: [0, 1], description: '0 Draft, 1 InReview; numeric candidate wire format remains under central review.' }
        newState: { type: integer, const: 0, description: 'Draft after the significant scope change.' }
    Problem:
      type: object
      required: [title, status]
      properties:
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        traceId: { type: string }
...
    PublishCommand:
      type: object
      additionalProperties: false
      required: [expectedContentVersion, expectedStateVersion]
      properties:
        expectedContentVersion: { type: integer, minimum: 0 }
        expectedStateVersion: { type: integer, minimum: 0 }
    RollbackDraftCommand:
      type: object
      additionalProperties: false
      required: [reason, expectedSourceStateVersion]
      properties:
        reason: { type: string, minLength: 1, maxLength: 2000 }
        expectedSourceStateVersion: { type: integer, minimum: 1 }
    InvalidateCommand:
      type: object
      additionalProperties: false
      required: [expectedContentVersion, expectedStateVersion, impactCode, reason, evidenceReference]
      properties:
        expectedContentVersion: { type: integer, minimum: 0 }
        expectedStateVersion: { type: integer, minimum: 0 }
        impactCode: { type: string, enum: [IdentityOrScope, CalendarTimeUnitOrQuantity, SourceMethodOrPolicy, ContentIntegrity] }
        reason: { type: string, minLength: 1, maxLength: 2000 }
        evidenceReference: { type: string, minLength: 1, maxLength: 2000 }
      description: Business impact must be independently verified behind the evidence reference. Actor and time are resolved by the server, not trusted from the body.